+++ classes/add_trigger.py
@@ -26,8 +26,13 @@ class addTrigger(wx.Dialog):
 
 		self.a=a
 
+		self.datastream_list2=[]
+		for i in datastream_list:
+			self.datastream_list2.append(i)
+		self.datastream_list2.append(_('None (always true)'))
+
 		wx.StaticText(panel, label=_('trigger'), pos=(10, 10))
-		self.trigger_select= wx.ComboBox(panel, choices=datastream_list, style=wx.CB_READONLY, size=(310, 32), pos=(10, 35))
+		self.trigger_select= wx.ComboBox(panel, choices=self.datastream_list2, style=wx.CB_READONLY, size=(310, 32), pos=(10, 35))
 		self.trigger_select.Bind(wx.EVT_COMBOBOX, self.onSelect)
 		wx.StaticText(panel, label=_('operator'), pos=(10, 70))
 		self.operator_select= wx.ComboBox(panel, choices=self.a.operators_list, style=wx.CB_READONLY, size=(310, 32), pos=(10, 95))
@@ -40,16 +45,21 @@ class addTrigger(wx.Dialog):
 		okBtn = wx.Button(panel, wx.ID_OK, pos=(180, 205))
 
 	def onSelect(self,e):
-		trigger=self.a.DataList[self.trigger_select.GetCurrentSelection()]
-		operators_valid_list=eval('self.a.'+trigger+'[7]')
-		new_list=[]
-		for i in operators_valid_list:
-			new_list.append(self.a.operators_list[i])
-		self.operator_select.Clear()
-		self.operator_select.AppendItems(new_list)
-		self.operator_select.SetSelection(0)
-
-		disable_field=eval('self.a.'+trigger+'[8]')
-		self.value.SetValue('')
-		if disable_field==1: self.value.Enable()
-		if disable_field==0: self.value.Disable()
+		if (self.trigger_select.GetCurrentSelection())+1==len(self.datastream_list2):
+			self.operator_select.Disable()
+			self.value.Disable()
+			wx.MessageBox(_('The actions of this trigger will always be executed.'), 'Info', wx.OK | wx.ICON_INFORMATION)
+		else:
+			trigger=self.a.DataList[self.trigger_select.GetCurrentSelection()]
+			operators_valid_list=eval('self.a.'+trigger+'[7]')
+			new_list=[]
+			for i in operators_valid_list:
+				new_list.append(self.a.operators_list[i])
+			self.operator_select.Enable()
+			self.operator_select.Clear()
+			self.operator_select.AppendItems(new_list)
+			self.operator_select.SetSelection(0)
+			disable_field=eval('self.a.'+trigger+'[8]')
+			self.value.SetValue('')
+			if disable_field==1: self.value.Enable()
+			if disable_field==0: self.value.Disable()
