+++ ctrl_actions.py
@@ -21,6 +21,9 @@ from classes.paths import Paths
 
 action=sys.argv[1]
 
+#see actions.py
+start_all_actions='18'
+
 paths=Paths()
 currentpath=paths.currentpath
 
@@ -29,17 +32,26 @@ conf=Conf()
 data=conf.get('ACTIONS', 'triggers')
 triggers=data.split('||')
 triggers.pop()
-
 for index,item in enumerate(triggers):
 	ii=item.split(',')
 	triggers[index]=ii
 
-tmp=''
+data=conf.get('ACTIONS', 'actions')
+trigger_actions=data.split('||')
+trigger_actions.pop()
+for index,item in enumerate(trigger_actions):
+	ii=item.split(',')
+	trigger_actions[index]=ii
 
+tmp=''
 #stop all
 if action=='0':
 	for index,item in enumerate(triggers):
-		tmp +='0,'
+		start_all=False
+		for i in trigger_actions:
+			if int(i[0])==index and i[1]==start_all_actions: start_all=True
+		if start_all==False: tmp +='0,'
+		if start_all==True: tmp +='1,'
 		tmp +=triggers[index][1]+','+triggers[index][2]+','+triggers[index][3]+'||'
 	conf.set('ACTIONS', 'triggers', tmp)
 	subprocess.Popen(['pkill', '-f', 'message.py'])
@@ -53,6 +65,6 @@ if action=='1':
 	conf.set('ACTIONS', 'triggers', tmp)
 
 subprocess.call(['pkill', '-f', 'monitoring.py'])
-subprocess.Popen(['python',currentpath+'/monitoring.py', '0'])
+subprocess.Popen(['python',currentpath+'/monitoring.py'])
 
 sys.exit()
